requirements.txt
python-dotenv
discord
datetime
asyncio